Move upgradesteps into a flat structure without v5 references
